+++ src/features/User/index.js
@@ -129,7 +129,7 @@ class User extends Component {
     return (
       <Content
         className={cx({ [styles.invalid]: safeUser.valid === false })}
-        cardExtra={safeUser.access && (
+        cardExtra={(
           <TooltipTrigger data={tooltip}>
             <Icon size="mini" className={styles.access} name={icon} />
           </TooltipTrigger>